Fix "format '%d' expects type 'int', but argument 6 has type 'size_t'" warning.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@27714 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libao2/ao_alsa.c
@@ -707,7 +707,7 @@ static int init(int rate_hz, int channels, int format, int flags)
       /* end setting sw-params */
 
       mp_msg(MSGT_AO,MSGL_V,"alsa: %d Hz/%d channels/%d bpf/%d bytes buffer/%s\n",
-	     ao_data.samplerate, ao_data.channels, bytes_per_sample, ao_data.buffersize,
+	     ao_data.samplerate, ao_data.channels, (int)bytes_per_sample, ao_data.buffersize,
 	     snd_pcm_format_description(alsa_format));
 
     } // end switch alsa_handler (spdif)
